World: Fix server health metrics for single-data point sets
The standard deviation is undefined (but could be set to zero) if we're measuring a single tick. Since this scenario doesn't make much sense, it's probably best to just throw here and be done with it.

This will likely never occur in practice unless someone sets the world server's tick interval to something impractical, but it has come up in the unit tests. Because the fix is trivial and has no relevant performance impact, let's just do the right thing here and handle it.