World: Improve the performance of the update loop
As a follow-up to #11, moving from a sleep-based to a timer-based loop seems to almost double the throughput.

Since the event loop sleeps internally anyway, this should be functionally identical and not cost (significantly) more CPU, so there's no reason to sleep manually here.

Unfortunately, testing timers is always icky, but degradation should be obvious if the target frame rate isn't reached (while idle).