Repo: Set package.path explicity to avoid OS-dependent lookups
By default, the system LUA_PATH is used to determine the lookup path. While this works with a default configuration (as evidenced by CI runs), it's not guaranteed to work. Since the general assumption (for now) is to use standard require with relative paths from the project root, it would be prudent to make sure the lookups always succeed.