Formats: Fix GRF extraction for UTF8-encoded file paths
If the name has been normalized before, as is the case when using the file entry to extract all files from a given GRF archive, the lookup would fail because the UTF8-encoded name is again multibyte-decoded, which yields incorrect results.

Instead of doing fancy checks, let's keep it simple and assume that this is the only situation to deal with. Checking both entries while performing the lookup is the most straightforward solution.