Client: Set a default window size even in fullscreen mode
It seems that the fullscreen mode on OSX is a bit odd: The runtime fails to show the window when no size has been specified, and then when it tries to toggle fullscreen mode it can't scale up the window (perhaps its size is initialised with zero?).

That aside, it's probably a good idea to set a default size anyway.