Client: Fix a glitch in the double-click camera reset logic
When releasing the mouse button and immediately clicking, the view is always reset - even if the button was pressed a long time ago (or, as in the test case, never). This is unintuitive, as the unspoken convention used by the client currently is that a keypress should be triggered on "press" and never "release".

Ideally, that'd be configurable but it's not that important right now.