Refactor: Eliminate viewport queries from the Renderer
The sole responsibility of the NativeClient module is to handle the top-level window, while the renderer should only care about its render target (in this case, the backing surface). Resize events are also handled by the window, or at least they should be (NYI), so this separation of concerns likely makes things easier.