Refactor: Streamline the GPU initialization module
It serves no useful purpose after the adapter has been obtained, so there's not much of a point in treating it like a proper library (that would've potentially needed to be moved to the runtime eventually).

In the context of this project, it's essentially the browser's top-level GPU class, although that one doesn't even have access to the instance.