GND: Implement a terrain geometry generation algorithm
Lightmaps and normals are still missing, but can easily be added later.

After much trial and error, the current approach seems to be the best in terms of both extensibility and performance. It's probably not exactly optimal still, but I'll accept a worst-case generation time of 200ms for schg_dun01 as "good enough" for now.