Refactor: Hide all WebGPU API calls behind facade classes
This might seem a bit questionable (and perhaps I'll end up reverting it), but ideally all external API calls would be hidden so that they can be unit tested more freely, where it makes sense.

The facade is a 1:1 mapping to the WebGPU API that currently exposes about 180 functions, a tiny subset of which are actually used by the renderer. That is to say, the maintenance overhead should be negligible as I don't plan to use most of the features.

There may be an opportunity to add event tracing/logging here, too.