Runtime: Extend the runtime to allow detecting test runs
Not sure about this, but the idea is to detect test runs in GPU code (to drop WebGPU API calls that won't work in the CI). It's a first step towards adding more testable instrumentation in the renderer.

If this turns out to be useful, it could be moved into the runtime itself, but for the time being it's just an experiment that doesn't warrant a dedicated runtime upgrade.