Cleanup: Remove the obsolete WebClient prototype
WebGPU is the future. I'll re-add the relevant bits and pieces later and/or re-implement them as needed... without the React/TypeScript madness or browser limitations.

Leaving  the placeholder mini map asset to be version-controlled, as it can't be backed up via GitHub gists and otherwise I'll forget where I put the thing...