Client: Implement a streamlined scene loading process
Streamlined compared to the ad-hoc prototype code... still not exactly a fully-fledged scene graph or anything like that.

It's still very much lacking in tests and leaks memory, but I'll get to that later. Things should be a bit easier with a standardized resource management API so I'm waiting for that (coming Soon ™).