Client: Rework the stress test to fix crashes on missing GND
As described in #244, Koko's classic map name table doesn't play nice with the latest kRO GRF (obviously). So as a stopgap measure, this fixes the incessant crashing by only attempting to load maps that exist in both the GRF and the map database.

Obviously, this isn't ideal since it skips all the Renewal formats.