GRF: Allow DecodeFileName to be used as a static method
The preallocation happens in the constructor, but that isn't really required in this case since the method doesn't depend on the decoded file table (or other internal data). Multithreading isn't an issue here either, so there's no reason to not share a single conversion buffer. In fact, enabling the use of this as a static method makes it more convenient for other decoders to use the conversion code... although it's questionable that they should need to in the first place.