Client: Add viewport dimensions to the per-scene uniform data
The time value isn't useful, IIRC it's a leftover from the original prototype that was based on the WebGPU tutorial... So let's replace it with something that the widget drawing pipeline actually needs - this way, the same uniform struct can be reused. There's plenty of unused space anyway due to the alignment requirements.