Client: Introduce a per-mesh uniform struct for widget transforms
Using the supported limits ensures that the device can always be obtained, and so it's reasonable to use this as a default. However, now that the widget rendering pipeline requires the use of dynamic uniform buffers, it could become  a problem. Instead of potentially causing invisible ( = hard-to-debug) problems down the line, hardcode it to 32 since that always works on my device.

Clearly, this isn't a general solution. See #272 for details on this issue.

The alignment function is an unused placeholder; leaving it in as another reminder. It's from the WebGPU tutorial so it should work.