Client: Streamline the texture generation to enable RML events
RML needs access to some of the texture info to create UI textures on-the-fly. This has to happen inside the C++ layer, invisible to LuaJIT, as otherwise all of the geometry and texture data would have to be copied around several times (at least twice on the CPU and then to VRAM).

Instead of all that silliness, the renderer can just receive the resources and later free them, reducing the number of copies to one (RML -> GPU buffer). The downside is that things get a little messy, but it's doable.