GND: Add a method to translate from grid to world coordinates
It's a bit slower than the code used in the terrain generation since it computes more stuff that simply isn't needed there, so I haven't reused it (even though that works just fine). May need to optimize later, but for now that doesn't matter as I don't have any performance-critical use cases in mind. The tests are painfully repetitive, so that may need some streamlining as well (deferred).