RSW: Add a surface region to the AnimatedWaterPlane class
The dimensions of the water plane are obvious for classic maps since it aligns perfectly with the ground mesh, but in Renewal there's a little more work required to properly align and place multiple planes correctly. The classic plane may be seen as a special case of the later (more general) approach, so the same tiling method can be used.