Client: Enable sorting meshes by material before rendering
This should come in handy once the materials use different pipelines each (that is to say, very soon).

I've started with three materials that correspond to the different pipeline configurations needed to implement key features:

* Ground meshes must support prebaked lightmap textures
* Water surfaces must support waveform sampling and texture cycling
* Everything else should just fall back to the default rendering pipeline

Actually implementing all of the above will take time, but materials are essential for adding things like that without headaches.