Perf: Avoid rebinding the camera uniform for every mesh
The "scenewide" uniform data is changing exactly once per frame (in UpdateScenewideUniformBuffer), so there's no point in adding redundant API calls to the render pass encoder for every mesh. It's possible the implementation would optimize them away, and there aren't enough meshes being rendered for it to become a problem right now, but such a potential footgun shouldn't be ignored.