Refactor: Streamline the use of localized ffi.string references
In this case, using "string" as the upvalue name overrides the builtin string library for that scope, which is a potential footgun and should be avoided. On the other hand, using "tostring" seems to work even if regular Lua strings are passed, but I'd consider it another footgun since code that expects to call the "regular" tostring function might want the cdata values printed, and either way ffi.string is unsafe so accidentally calling it could be causing a SEGFAULT - let's not do that.