Refactor: Move texture bind groups to material instances
In order to support uniform material properties, control of the uniform buffer and layout needs to be contained in the material instance. It's not enough to create just one uniform buffer, though, since each diffuse texture would have to be bound using its own bind group (just as it is now).

The easiest way to untangle the current method where texture allocate their own bind groups is to let the material instance allocate it and handle all the texture binding as well.

For UI texture things will get more complicated since RML has no concept of the material system. For now, leaving the old method in to support the UI rendering pipeline seems fine, although the UI will have to be moved to the material system eventually.