Client: Add a Workaround for the broken material inheritance
Unfortunately, the new OOP library has a tiny issue:

If in the derived class, a  constructor wasn't set (because it was meant to be inherited), then class(...) will generate a default constructor, which prevents extend(...) from properly looking up the prototype's implementation. In this case, that means material properties - such as the displayName - can't be looked up. A proper fix will need to be included in the runtime, or hacked in here (but let's not go there).