RSW: Fix an oversight in the water vertex culling strategy
This was even present in the JS version, but somehow I forgot to add it here. Needless to say, culling water plane segments that could become visible at the height of their wave cycle leads to visual glitches as that part of the surface will appear to have quad-shaped holes in it.