Tools: Create a script to analyze RSW and GND water planes
Had to jump through some hoops to avoid GBs of memory usage, but now it seems to be fine (if a little wasteful, taking maybe 300 MB or so in total). Could be optimized further by analysing each resource immediately and then discarding (or reusing) the memory, but it really doesn't matter here.