Refactor: Enable arbitrary number formats for performance metrics
The code originally assumes time (in milliseconds), which doesn't allow for displaying RAM usage or other metrics. Since time is a sensible default, overriding the format for specific items seems generalized enough for now.

A better approach might be to allow passing formatter functions and using tables instead of varargs (to be order-independent). But it's probably not needed, so defer until there's a reason to generalize further.