Client: Increase the maximum GPU buffer size to 256 MB
This is the default WebGPU limit, so it should be safe to use. The existing one is too small to render large maps with normal visualization enabled (such as schg_dun01 or nif_fild01) - there's just too many vertices since each vertex normal necessitates three more for the visualization itself.