Core: Add a modified version of the fog parameters decoder
The code is ancient; it was part of the now-archived RagnarokTools repository. I didn't change too much about it so it's still questionable, but I suppose that's really not a problem here. Only minor adjustments were made to make use of the Color metatype and to make the code more testable.