Tests: Fix the checksum tests for generated GAT images
The tests actually try to assert that the generated PNG bytes are identical, but this isn't the case when running them on M1-based macOS systems. It also doesn't really make sense; only the image contents should be validated and not the on-disk representation.

The fact that this worked at all is likely a happy accident.