Tools: Rework the GAT heightmap image generation script
The previous approach to normalization is nonsensical and fails when encountering certain values, including those used by the test fixtures. This manifests differently on the M1 runner (using ARM64), which is why the CI kept failing even after the last few fixes. Not sure if this normalization method is the best choice, as the contrast is quite low in certain maps (e.g., prontera). But at least it shouldn't generate NaN/invalid results.