Client: Implement screenshot captures on demand
For now they're blocking, since the time taken isn't that long. It does skip a few frames, but offloading the encoding step would be more complex than just doing it on the main thread and eating the dropped frames.

It can also timeout the surface texture and crash the renderer, but that's a separate problem and will be tackled soon (TM).