Client: Implement caching for mesh geometry buffer sizes
This can be used to compute the actually required buffer size after the first load, and then pre-allocate the buffers via table.new as is currently already happening (but with a perfectly matching buffer size, which avoids under- or over-allocation).

Obviously this isn't the best caching mechanism, but it's better than nothing and might already be enough to speed up loading times (at least for now).