Tools: Add a utility script for analyzing GRF paths
This isn't the most useful right now, but I can imagine wanting to add more analysis steps later on. At that point it should likely be moved to the FileAnalyzer module as well.

Right now it's sufficient to determine the path lengths since it allowed tinkering with some optimizations that trade off space usage for decoding time; if compiling the entries as fixed-size blocks, can reduce the decoding time by 30 ms (from 70), but disk use increases by a lot so I'll start with a different approach.