Client: Fix an unintended conversion in the terrain shader
This error was concealed by a recent change to wgpu, where it would automatically convert unsigned integers to floats when it deems it "safe" to do so. Previously, it would've given a validation error here because the hardcoded numbers aren't compiled to floats unless explicitly marked as such.

Apparently, this hidden conversion broke the discard logic, leading to black surfaces being discarded in at least one case (c_tower1). This is because the division doesn't have enough precision if using integers. Being explicit fixes the problem.