Client: Create a demo scene for mesh instancing tests
Right now it doesn't use hardware instancing, which is why the performance is awful. I want to see how big of a difference there will be, so let's use this as a primitive benchmark:

If the renderer can do 2500 cubes before dropping below 60 FPS, it should do a lot better with instancing enabled - but also a lot worse if those meshes are actually complex and textured.