Client: Fix the alpha blending for semi-translucent water surfaces
After some investigation, it's become clear that the reference client seems to incorrectly mix sRGB and linear-space colors. While this is only a theory, it would explain why the contrast correction was needed in the first place; it makes up for the underexposure introduced by the lack of gamma-decompression steps after loading sRGB-encoded textures. Indeed, screen blending is used for this exact purpose in digital photo editing (apparently).

In order to reproduce the resulting outputs, neither fully-sRGB nor fully-linearized computations will yield accurate blends while the underlying texture view is in sRGB format. WebGPU automatically converts the fragment colors - first to sRGB and then back to linear as part of the blend stage. Unfortunately, the color space used is actually neither of the two by this point, and so the results can't possibly match expectations.

Originally, sRGB was selected because it's the preferred surface format. Using a different one grants full control of the color space used and eliminates the undesirable conversion, but may incur a performance hit. However, conversions between RGBA and sRGB should be hardware-accelerated on any reasonably modern device. Switching to RGBA also simplifies the shader logic as the added linearization step is no longer needed. So overall it's probably a net gain?