Client: Improve the accuracy of downsampled texture colors
It seems that some error correction is applied by the fixed-function pipeline when operating on the downsampled (reduced-color) textures? Merely truncating the colors, as previously done, yields visible artifacts with an accumulating error consistent with the dropped precision (1 per 32 for bpp = 5). The base colors would have to be either uint8 or floats, which carry more information after all. The results are otherwise fairly exact, at least on the few maps I tested. This may also apply to lightmaps? (TBD)