Client: Create a standalone Win32 "client" application
Synchronizing the Lua runtime just to use a bunch of native libraries is bit tedious. Also, it's much slower (obviously...). That isn't much of a problem right now, but I'd like to see how much of a difference it makes while decoding and analyzing file formats en masse.

For now, the idea is to eliminate dependencies and see how that works out. If needed it would be fairly easy to pull in the same libraries that are currently used (GLFW, WebGPU, etc.). Building them is annoying so I'd rather not do it just yet.

(If the experiment works out, can port to Linux/macOS later)