Client: Port WebGPU renderer to the v0.0.22 glfw API
**WARNING: THE RENDERER IS COMPLETELY BROKEN!**

glfw3wgpu doesn't work in the latest version, because it uses outdated WGPU structures and fails to create the surface from a native window handle. This then yields a mountain of rust stack unwinding errors after wgpu-native dies with "invalid surface".

I'm not sure if it's worth fixing here since the tools and renderer work fine on v0.0.020 (except for Wayland - oh, well). Will have to wait and see what the best course of action is. It might require FFI access to GLFW and X11/Wayland/Cocoa/Metal APIs?