Win32: Remove unnecessary redraw window flags to reduce WM_PAINT messages
These flags cause the window to receive additional WM_PAINT messages as Windows wants the app to redraw each invalidated rectangle. But in the GDI render loop, the program invalidates the entire client rectangle anyway, so these aren't useful and spam the message queue with repaint events when resizing the window.