Win32: Skip draw calls while the main window is minimized
Although the thread seems to run less often, an unavoidable minimization delay still skews the recorded metric. This then leads to visible spikes in the XY plot, and is of course wasteful. The ideal UX is somewhat unclear, so just stop drawing entirely as that is by far the most expensive operation currently.